Differentiating the auth and non-auth LTI params schema
 * We now have three schema:
   * LaunchParamsSchema - Checks required for LTI
   * LaunchParamsURLConfiguredSchema - The above + a URL decoding step
   * LaunchParamsAuthSchema - Contains only checks for oath
 * The auth and others are separated (along with tests)
