+++ lms/validation/__init__.py
@@ -48,6 +48,7 @@ from lms.validation._canvas import *
 from lms.validation._exceptions import *
 from lms.validation._lis_result_sourcedid import LISResultSourcedIdSchema
 from lms.validation._module_item_configuration import ConfigureModuleItemSchema
+from lms.validation._lti_launch_params import *
 
 __all__ = (
     "APIRecordSpeedgraderSchema",
@@ -57,6 +58,7 @@ __all__ = (
     "ConfigureModuleItemSchema",
     "CanvasListFilesResponseSchema",
     "CanvasPublicURLResponseSchema",
+    "LaunchParamsURLConfiguredSchema",
     "ValidationError",
 )
