+++ lms/validation/authentication/_launch_params.py
@@ -1,5 +1,4 @@
 """Schema for validating LTI launch params."""
-from urllib.parse import unquote
 
 import marshmallow
 
@@ -7,10 +6,10 @@ from lms.services import LTILaunchVerificationError
 from lms.validation._helpers import PyramidRequestSchema
 from lms.values import LTIUser
 
-__all__ = ("LaunchParamsSchema", "URLConfiguredLaunchParamsSchema")
+__all__ = ("LaunchParamsAuthSchema",)
 
 
-class LaunchParamsSchema(PyramidRequestSchema):
+class LaunchParamsAuthSchema(PyramidRequestSchema):
     """
     Schema for LTI launch params.
 
@@ -18,13 +17,13 @@ class LaunchParamsSchema(PyramidRequestSchema):
 
         >>> from webargs.pyramidparser import parser
         >>>
-        >>> schema = LaunchParamsSchema(request)
+        >>> schema = LaunchParamsAuthSchema(request)
         >>> parsed_params = parser.parse(schema, request, locations=["form"])
 
     Or to verify the request and get an :class:`~lms.values.LTIUser`
     from the request's params::
 
-        >>> schema = LaunchParamsSchema(request)
+        >>> schema = LaunchParamsAuthSchema(request)
         >>> schema.lti_user()
         LTIUser(user_id='...', ...)
     """
@@ -71,26 +70,3 @@ class LaunchParamsSchema(PyramidRequestSchema):
             raise marshmallow.ValidationError("Invalid OAuth 1 signature.") from err
 
 
-class URLConfiguredLaunchParamsSchema(PyramidRequestSchema):
-    """
-    Schema for an "URL-configured" Basic LTI Launch.
-
-    An URL-configured launch is one where the content URL is provided by a "url"
-    launch param.
-    """
-
-    url = marshmallow.fields.Str(required=True)
-
-    @marshmallow.post_load
-    def _decode_url(self, _data, **_kwargs):  # pylint:disable=no-self-use
-        # Work around a bug in Canvas's handling of LTI Launch URLs in
-        # SpeedGrader launches. In that context, query params get
-        # doubly-encoded. This is worked around by detecting when this has
-        # happened and decoding the URL a second time.
-        #
-        # See https://github.com/instructure/canvas-lms/issues/1486
-        url = _data["url"]
-        if url.lower().startswith("http%3a") or url.lower().startswith("https%3a"):
-            url = unquote(url)
-            _data["url"] = url
-        return _data
